backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.28.1
cloudinary==1.41.0
email-validator==2.2.0
python-multipart==0.0.9
passlib==1.7.4
python-jose==3.3.0
qrcode==7.4.2
Pillow==10.4.0
